+++ src/main/java/wooteco/subway/domain/path/Fare.java
@@ -1,4 +1,4 @@
-package wooteco.subway.domain;
+package wooteco.subway.domain.path;
 
 public class Fare {
