+++ src/main/java/wooteco/subway/dto/error/ErrorResponse.java
@@ -1,4 +1,4 @@
-package wooteco.subway.dto;
+package wooteco.subway.dto.error;
 
 public class ErrorResponse {
