+++ src/main/java/wooteco/subway/dto/path/PathResponse.java
@@ -1,4 +1,4 @@
-package wooteco.subway.dto;
+package wooteco.subway.dto.path;
 
 import java.util.List;
 import wooteco.subway.dto.station.StationResponse;
